Refactored "Teams" page components. - brought components to a common structure - dropped redundant component - cleaned up global styles a bit
