+++ components/page/teams/TeamList/components/TeamGridView/index.ts
@@ -1,0 +1,1 @@
+export * from './TeamGridView';
